In browse, move ModelGroup and associated components to a separate file
