+++ frontend/src/metabase/browse/components/ModelGroup.tsx
@@ -1,0 +1,135 @@
+import type {
+  Card,
+  CollectionEssentials,
+  SearchResult,
+} from "metabase-types/api";
+import * as Urls from "metabase/lib/urls";
+
+import Link from "metabase/core/components/Link";
+import Search from "metabase/entities/search";
+import { useDispatch } from "metabase/lib/redux";
+
+import { Box, Group, Icon, Text, Title } from "metabase/ui";
+
+import { color } from "metabase/lib/colors";
+import { getCollectionName } from "../utils";
+import {
+  CollectionHeaderContainer,
+  CollectionHeaderGroup,
+  CollectionHeaderLink,
+  ModelCard,
+  MultilineEllipsified,
+} from "./BrowseModels.styled";
+import { LastEdited } from "./LastEdited";
+
+export const ModelGroup = ({
+  models,
+  localeCode,
+}: {
+  models: SearchResult[];
+  localeCode: string | undefined;
+}) => {
+  const sortedModels = models.sort((a, b) => {
+    if (!a.name && b.name) {
+      return 1;
+    }
+    if (a.name && !b.name) {
+      return -1;
+    }
+    if (!a.name && !b.name) {
+      return 0;
+    }
+    const nameA = a.name.toLowerCase();
+    const nameB = b.name.toLowerCase();
+    return nameA.localeCompare(nameB, localeCode);
+  });
+  const collection = models[0].collection;
+
+  /** This id is used by aria-labelledby */
+  const collectionHtmlId = `collection-${collection.id}`;
+
+  // TODO: Check padding above the collection header
+  return (
+    <>
+      <CollectionHeader
+        collection={collection}
+        key={collectionHtmlId}
+        id={collectionHtmlId}
+      />
+      {sortedModels.map(model => (
+        <ModelCell
+          model={model}
+          collectionHtmlId={collectionHtmlId}
+          key={`model-${model.id}`}
+        />
+      ))}
+    </>
+  );
+};
+
+interface ModelCellProps {
+  model: SearchResult;
+  collectionHtmlId: string;
+}
+
+const ModelCell = ({ model, collectionHtmlId }: ModelCellProps) => {
+  const headingId = `heading-for-model-${model.id}`;
+
+  const lastEditorFullName =
+    model.last_editor_common_name ?? model.creator_common_name;
+  const timestamp = model.last_edited_at ?? model.created_at ?? "";
+
+  return (
+    <Link
+      aria-labelledby={`${collectionHtmlId} ${headingId}`}
+      key={model.id}
+      to={Urls.model(model as unknown as Partial<Card>)}
+    >
+      <ModelCard>
+        <Box mb="auto">
+          <Icon name="model" size={20} color={color("brand")} />
+        </Box>
+        <Title mb=".25rem" size="1rem">
+          <MultilineEllipsified tooltipMaxWidth="20rem" id={headingId}>
+            {model.name}
+          </MultilineEllipsified>
+        </Title>
+        <LastEdited editorFullName={lastEditorFullName} timestamp={timestamp} />
+      </ModelCard>
+    </Link>
+  );
+};
+
+const CollectionHeader = ({
+  collection,
+  id,
+}: {
+  collection: CollectionEssentials;
+  id: string;
+}) => {
+  const dispatch = useDispatch();
+  const wrappable = { ...collection, model: "collection" };
+  const wrappedCollection = Search.wrapEntity(wrappable, dispatch);
+  const icon = wrappedCollection.getIcon();
+
+  return (
+    <CollectionHeaderContainer
+      id={id}
+      role="heading"
+      pt={"1rem"}
+      mr="1rem"
+      align="center"
+    >
+      <CollectionHeaderGroup grow noWrap>
+        <CollectionHeaderLink to={Urls.collection(collection)}>
+          <Group spacing=".25rem">
+            <Icon {...icon} />
+            <Text weight="bold" color="text-dark">
+              {getCollectionName(collection)}
+            </Text>
+          </Group>
+        </CollectionHeaderLink>
+      </CollectionHeaderGroup>
+    </CollectionHeaderContainer>
+  );
+};
